Add files via upload
A more refined code versus the first attempt. I was on the right track, but game up that this line of thinking with the first attempt. Also added in a check to change the characters to lower case and comparing the original and new string to count how many vowels were deleted/removed.